requirements.txt
pandas==2.0.2
pyarrow==12.0.1
polars==0.18.4
pushbullet.py~=0.12.0
pycoingecko~=3.1.0
python-binance
ccxt==3.1.54
websocket-client==1.6.1
numpy==1.25.0
plotly==5.15.0
joblib~=1.2.0
pytz~=2023.3
mlxtend
imblearn
scikit-learn~=1.2.2
scikit-learn-intelex